Surface the ds64 sample count, widen data offsets, drop duplicate chunks
Three smaller fidelity gaps.

`ds64` carries the sample-frame count for RF64 files, where `fact` carries it
for plain RIFF, but it was parsed and thrown away. Both now reach the caller,
through `ds64_sample_count` and a `sample_count()` that reads whichever the
file has.

`data_offset` and `data_length` were `usize`, which caps an RF64 file at 4 GB
on a 32-bit target, the exact case RF64 exists for. Both are `u64` now. The
frame count stays `usize` because it indexes buffers in memory, and a declared
length past what the target can address is clamped rather than overflowing.

A second `fmt `, `data` or `fact` chunk is now dropped rather than captured.
None of the three is valid in a WAVE file to begin with, and capturing them
would be worse than useless: all three ids are reserved on the write side, so a
caller feeding the chunks back would be handed something the writer refuses.
For `data` there is also no bound on how much audio a second one would pull
into memory just to hand back.

+++ tests/roundtrip.rs
@@ -688,7 +688,7 @@ fn rf64_roundtrip() {
     assert_eq!(reader.sample_format(), Some(SampleFormat::F32));
     assert_eq!(reader.channels(), channels);
     assert_eq!(reader.frames(), frames);
-    assert_eq!(reader.params().data_length, data_bytes as usize);
+    assert_eq!(reader.params().data_length, data_bytes);
     let restored = reader.read_all_to_float::<f32>().unwrap();
     assert_eq!(restored.frames(), frames);
     for frame in 0..frames {
@@ -1518,7 +1518,7 @@ fn rf64_truncate_updates_the_ds64_sizes() {
     cursor.set_position(0);
     let reader = WavReader::new(cursor).unwrap();
     assert_eq!(reader.frames(), 12);
-    assert_eq!(reader.params().data_length, 12 * channels * 2);
+    assert_eq!(reader.params().data_length, (12 * channels * 2) as u64);
 }
 
 #[test]
